Refactor: modularize popup.js into session.js and ui.js
