+++ src/popup/popup.html
@@ -56,6 +56,6 @@
 
   <div id="floating-warning" class="floating-warning">Please select at least one tab before starting.</div>
 
-  <script src="popup.js"></script>
+  <script type="module" src="popup.js"></script>
 </body>
 </html>
