+++ src/popup/ui.js
@@ -1,0 +1,100 @@
+function updateResumeButtonToPause(pauseBtn) {
+  pauseBtn.textContent = "⏸ PAUSE";
+  pauseBtn.style.backgroundColor = "#facc15";
+}
+
+function updatePauseButtonToResume(pauseBtn) {
+  pauseBtn.textContent = "▶ RESUME";
+  pauseBtn.style.backgroundColor = "#d1d5db";
+}
+
+function updateUIForPausedState(pauseBtn, sessionInputs) {
+  updatePauseButtonToResume(pauseBtn);
+  if (sessionInputs) {
+    sessionInputs.forEach(input => input.disabled = true);
+  }
+}
+
+function disableTimerInputs(disable, hrInput, minInput, secInput) {
+  hrInput.disabled = minInput.disabled = secInput.disabled = disable;
+}
+
+function setTabTogglesVisible(visible) {
+  const items = document.querySelectorAll(".item");
+  items.forEach(item => {
+    const checkbox = item.querySelector("input[type='checkbox']");
+    if (checkbox) checkbox.style.display = visible ? "inline-block" : "none";
+    item.classList.toggle("no-blur", !visible);
+  });
+}
+
+
+function showFloatingWarning(message = "Please select at least one tab before starting.") {
+  const warning = document.getElementById("floating-warning");
+  warning.textContent = message;
+  warning.classList.add("show");
+
+  const removeWarning = () => {
+    warning.classList.remove("show");
+    document.removeEventListener("click", removeWarning, true);
+    document.removeEventListener("keydown", removeWarning, true);
+  };
+
+  // Remove on any interaction or after 300ms
+  setTimeout(removeWarning, 2000);
+  document.addEventListener("click", removeWarning, true);
+  document.addEventListener("keydown", removeWarning, true);
+}
+
+function renderTabList(tabs) {
+  const container = document.getElementById("tab-list");
+  container.innerHTML = ""; // Clear previous content
+
+  tabs.forEach((tab) => {
+    const div = document.createElement("div");
+    div.className = "item";
+
+    const domain = new URL(tab.url).hostname;
+
+    const checkbox = document.createElement("input");
+    checkbox.type = "checkbox";
+    checkbox.value = domain;
+
+    const labelContainer = document.createElement("div");
+    labelContainer.className = "label-container";
+
+    const favicon = document.createElement("img");
+    favicon.className = "favicon";
+    favicon.src = tab.favIconUrl || "";
+    favicon.onerror = () => {
+      favicon.onerror = () => {
+        favicon.src = "/resources/icons/default_icon.png";
+      };
+      favicon.src = `https://www.google.com/s2/favicons?sz=64&domain=${domain}`;
+    };
+
+    const titleSpan = document.createElement("span");
+    titleSpan.className = "title";
+    const fullTitle = tab.title || tab.url;
+    titleSpan.textContent = fullTitle;
+    titleSpan.title = fullTitle;
+
+    labelContainer.appendChild(favicon);
+    labelContainer.appendChild(titleSpan);
+
+    div.appendChild(labelContainer);
+    div.appendChild(checkbox);
+
+    container.appendChild(div);
+  });
+}
+
+export {
+  renderTabList,
+  updateUIForPausedState,
+  disableTimerInputs,
+  setTabTogglesVisible,
+  updatePauseButtonToResume,
+  updateResumeButtonToPause,
+  showFloatingWarning
+};
